Packages related to password validation have been moved to the service layer.
