+++ service/src/main/java/com/epam/rft/atsy/service/passwordchange/validation/impl/PasswordUniqueRule.java
@@ -1,8 +1,8 @@
-package com.epam.rft.atsy.web.passwordchange.validation.impl;
+package com.epam.rft.atsy.service.passwordchange.validation.impl;
 
 import com.epam.rft.atsy.service.PasswordChangeService;
 import com.epam.rft.atsy.service.domain.PasswordChangeDTO;
-import com.epam.rft.atsy.web.passwordchange.validation.PasswordValidationRule;
+import com.epam.rft.atsy.service.passwordchange.validation.PasswordValidationRule;
 import com.epam.rft.atsy.web.security.UserDetailsAdapter;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
